+++ app/src/main/java/com/activesince93/playassetdeliverydemo/helper/Utils.java
@@ -1,0 +1,38 @@
+package com.activesince93.playassetdeliverydemo.helper;
+
+import android.content.Context;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
+import android.widget.Toast;
+
+public class Utils {
+    /**
+     * This method will check user is Connected to Internet or not
+     *
+     * @return true: Boolean if User is connected to Internet, false otherwise
+     */
+    public static boolean isInternetConnected(Context context) {
+        ConnectivityManager cm = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+        NetworkInfo activeNetwork = cm.getActiveNetworkInfo();
+        if (activeNetwork != null) { // connected to the internet
+            if (activeNetwork.getType() == ConnectivityManager.TYPE_WIFI) {
+                return true;
+
+            } else if (activeNetwork.getType() == ConnectivityManager.TYPE_MOBILE) {
+                // connected to the mobile provider's data plan
+                return true;
+            }
+        } else {
+            // not connected to the internet
+            return false;
+        }
+        return false;
+    }
+
+    /**
+     * This method is used to show toast
+     */
+    public static void showToast(Context context, String msg) {
+        Toast.makeText(context, msg, Toast.LENGTH_SHORT).show();
+    }
+}
