--- app/src/main/java/com/example/playassetdeliverydemo/ui/VideoPlayerActivity.java
@@ -1,122 +1,0 @@
-package com.example.playassetdeliverydemo.ui;
-
-import android.annotation.SuppressLint;
-import android.net.Uri;
-import android.os.Bundle;
-import android.view.View;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import com.example.playassetdeliverydemo.R;
-import com.google.android.exoplayer2.MediaItem;
-import com.google.android.exoplayer2.SimpleExoPlayer;
-import com.google.android.exoplayer2.ui.PlayerView;
-import com.google.android.exoplayer2.util.Util;
-
-public class VideoPlayerActivity extends AppCompatActivity {
-
-    private PlayerView playerView;
-    private SimpleExoPlayer player;
-    private boolean playWhenReady = true;
-    private int currentWindow = 0;
-    private long playbackPosition = 0;
-    private Uri uri;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_video_player);
-        initViews();
-    }
-
-    /**
-     * This method is used to initialize views
-     */
-    private void initViews() {
-        playerView = findViewById(R.id.playerView);
-        uri = getIntent().getData();
-    }
-
-    /**
-     * lifecycle method to initialize ExoPlayer
-     */
-    @Override
-    public void onStart() {
-        super.onStart();
-        if (Util.SDK_INT >= 24) {
-            initializePlayer();
-        }
-    }
-
-    /**
-     * lifecycle method to initialize ExoPlayer
-     */
-    @Override
-    public void onResume() {
-        super.onResume();
-        hideSystemUi();
-        if ((Util.SDK_INT < 24 || player == null)) {
-            initializePlayer();
-        }
-    }
-
-    /**
-     * lifecycle method to release ExoPlayer
-     */
-    @Override
-    public void onPause() {
-        super.onPause();
-        if (Util.SDK_INT < 24) {
-            releasePlayer();
-        }
-    }
-
-    /**
-     * lifecycle method to release ExoPlayer
-     */
-    @Override
-    public void onStop() {
-        super.onStop();
-        if (Util.SDK_INT >= 24) {
-            releasePlayer();
-        }
-    }
-
-    /**
-     * Method to initialize ExoPlayer
-     */
-    private void initializePlayer() {
-        if (player == null) {
-            player = new SimpleExoPlayer.Builder(this).build();
-            playerView.setPlayer(player);
-            MediaItem mediaItem = MediaItem.fromUri(uri);
-            player.setMediaItem(mediaItem);
-        }
-    }
-
-    /**
-     * Method to release ExoPlayer
-     */
-    private void releasePlayer() {
-        if (player != null) {
-            playWhenReady = player.getPlayWhenReady();
-            playbackPosition = player.getCurrentPosition();
-            currentWindow = player.getCurrentWindowIndex();
-            player.release();
-            player = null;
-        }
-    }
-
-    /**
-     * Method to hide System UI & show ExoPlayer in FullScreen
-     */
-    @SuppressLint("InlinedApi")
-    private void hideSystemUi() {
-        playerView.setSystemUiVisibility(View.SYSTEM_UI_FLAG_LOW_PROFILE
-                | View.SYSTEM_UI_FLAG_FULLSCREEN
-                | View.SYSTEM_UI_FLAG_LAYOUT_STABLE
-                | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
-                | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
-                | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION);
-    }
-}
