+++ app/src/main/res/layout/activity_main.xml
@@ -8,10 +8,19 @@
     tools:context=".ui.MainActivity">
 
     <Button
+        android:id="@+id/btn_it"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textAllCaps="false"
+        android:visibility="gone"
+        android:text="[NOT WORKING] Install-Time" />
+
+    <Button
         android:id="@+id/btn_ff"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:textAllCaps="false"
+        android:visibility="gone"
         android:text="Fast-Follow" />
 
     <Button
@@ -36,5 +45,11 @@
         android:visibility="gone"
         android:textAllCaps="false"
         android:text="Play Audio" />
+
+    <TextView
+        android:id="@+id/txt_download_status"
+        android:layout_width="wrap_content"
+        android:text="ASSET STATUS"
+        android:layout_height="wrap_content" />
 
 </LinearLayout>
